refactor(update): rename and move files to more sensible things
