+++ bin/upem.js
@@ -1,13 +1,13 @@
 #!/usr/bin/env node
 const path = require('path')
-const uppity = require('../src/uppity')
+const up = require('../src')
 const readFromStream = require('../src/readFromStream')
 
 const PACKAGE_FILE_NAME = path.join(process.cwd(), 'package.json')
 
 readFromStream()
   .then(pOutdatedObject => {
-    const lResult = uppity(PACKAGE_FILE_NAME, pOutdatedObject)
+    const lResult = up(PACKAGE_FILE_NAME, pOutdatedObject)
 
     if (lResult.OK) {
       process.stdout.write(lResult.message)
