+++ test/index.spec.js
@@ -1,86 +1,72 @@
+const fs = require('fs')
+const path = require('path')
 const up = require('../src')
-const DEPS_FIXTURE = {
-  'not-outdated': '1.0.0',
-  'outdated-one': '2.0.0',
-  'outdated-possibly-pinned': '3.1.4'
-}
-const DEPS_UPDATED_FIXTURE = {
-  'not-outdated': '1.0.0',
-  'outdated-one': '3.0.2',
-  'outdated-possibly-pinned': '4.1.1'
-}
-const OUTDATED_FIXTURE = {
-  'outdated-one': {
-    'current': '2.0.0',
-    'wanted': '2.0.1',
-    'latest': '3.0.2',
-    'location': 'node_modules/not-outdated'
-  },
-  'outdated-possibly-pinned': {
-    'current': '3.1.4',
-    'wanted': '3.1.4',
-    'latest': '4.1.1',
-    'location': 'node_modules/outdated-possibly-pinned'
-  }
-}
-describe('#updateDeps', () => {
-  test('empty deps, no outdated yield input', () => {
-    expect(up.updateDeps({}, [])).toEqual({})
+
+describe('#upppity', () => {
+  test('non-existing package.json errors', () => {
+    const lResult = up('thisfiledoesnotexist', '')
+    expect(lResult.OK).toEqual(false)
+    expect(lResult.message).toContain(`Up'em encountered a hitch:`)
   })
-  test('deps, no outdated yield input', () => {
-    expect(up.updateDeps(DEPS_FIXTURE, {})).toEqual(DEPS_FIXTURE)
+
+  test('empty string dependency JSON yields "nothing to update"', () => {
+    const lResult = up(path.join(__dirname, 'package-in.json'), '')
+    expect(lResult.OK).toEqual(true)
+    expect(lResult.message).toContain(`Up'em says: Everything seems to be up to date already.`)
   })
-  test('deps, no outdated yield input', () => {
-    expect(up.updateDeps(DEPS_FIXTURE, {})).toEqual(DEPS_FIXTURE)
+
+  test('{} dependency JSON yields "nothing to update"', () => {
+    const lResult = up(path.join(__dirname, 'package-in.json'), '{}')
+    expect(lResult.OK).toEqual(true)
+    expect(lResult.message).toContain(`Up'em says: Everything seems to be up to date already.`)
   })
-  test('deps, outdated yields updated deps', () => {
-    expect(up.updateDeps(DEPS_FIXTURE, OUTDATED_FIXTURE)).toEqual(DEPS_UPDATED_FIXTURE)
+
+  test(`read only package.json yields a 'can't update`, () => {
+    const OUTDATED_JSON = `
+    {
+      "lodash.assign": {
+        "current": "4.2.0",
+        "wanted": "4.2.0",
+        "latest": "4.2.1",
+        "location": "node_modules/lodash.assign"
+      }
+    }
+    `
+    const READONLY_INPUT_FILENAME = path.join(__dirname, 'package-in-readonly.json')
+    fs.chmodSync(READONLY_INPUT_FILENAME, '400')
+    const lResult = up(READONLY_INPUT_FILENAME, OUTDATED_JSON)
+    expect(lResult.OK).toEqual(false)
+    expect(lResult.message).toContain(`Up'em encountered a hitch when updating package.json:`)
+  })
+
+  test('if upem.donotup encompasses the outdated object yields that in a message', () => {
+    const OUTDATED_JSON = `
+    {
+      "ts-jest": {
+        "current": "1.8.2",
+        "wanted": "1.8.2",
+        "latest": "2.0.0",
+        "location": "node_modules/ts-jest"
+      }
+    }
+    `
+    const lResult = up(path.join(__dirname, 'package-in.json'), OUTDATED_JSON)
+    expect(lResult.OK).toEqual(true)
+    expect(lResult.message).toContain(`Up'em says: Everything not in 'upem.donotup' seems to be up to date already.`)
+  })
+
+  test('happy day: dependencies updated with stuff in an outdated.json', () => {
+    const OUTDATED_JSON = fs.readFileSync(path.join(__dirname, 'outdated.json'))
+    const INPUT_FILENAME = path.join(__dirname, 'package-in.json')
+    const OUTPUT_FILENAME = path.join(__dirname, 'tmp_package-out.json')
+    const FIXTURE_FILENAME = path.join(__dirname, 'package-out.json')
+
+    const lResult = up(INPUT_FILENAME, OUTDATED_JSON, OUTPUT_FILENAME)
+
+    expect(lResult.OK).toEqual(true)
+    expect(lResult.message).toContain(`Up'em just updated all dependencies in package.json to latest`)
+    expect(JSON.parse(fs.readFileSync(OUTPUT_FILENAME))).toEqual(JSON.parse(fs.readFileSync(FIXTURE_FILENAME)))
   })
 })
 
-describe('#updateAllDeps', () => {
-  test('empty package.json, no outdated yield input', () => {
-    expect(up.updateAllDeps({}, {})).toEqual({})
-    expect(up.updateAllDeps({})).toEqual({})
-  })
-
-  test('empty package.json, several outdated yield input', () => {
-    expect(up.updateAllDeps({}, ['aap', 'noot', 'mies'])).toEqual({})
-  })
-
-  test('real package.json, several outdated yield updated output', () => {
-    expect(
-      up.updateAllDeps(
-        require('./package-in.json'),
-        require('./outdated-filtered.json')
-      )
-    ).toEqual(require('./package-out.json'))
-  })
-})
-
-describe('#filterOutdatedPackages', () => {
-  test('empty outdated + empty package => empty outdated', () => {
-    expect(
-      up.filterOutdatedPackages({}, {})
-    ).toEqual({})
-  })
-
-  test('empty outdated + package => empty outdated', () => {
-    expect(
-      up.filterOutdatedPackages({}, require('./package-in.json'))
-    ).toEqual({})
-  })
-
-  test('outdated + package wit upem.donotup => outdated without the upem.donotup', () => {
-    expect(
-      up.filterOutdatedPackages(require('./outdated.json'), require('./package-in.json'))
-    ).toEqual(require('./outdated-filtered.json'))
-  })
-
-  test('outdated + package wit upem.donotup => outdated without the upem.donotup', () => {
-    expect(
-      up.filterOutdatedPackages(require('./outdated.json'), require('./package-in-without-upem-donotup.json'))
-    ).toEqual(require('./outdated.json'))
-  })
-})
 /* global describe, test, expect */
